fix(ast): rebase reconciliation — retire an obsolete mutation-guard test, split test_ast_parser.py under the 300-line cap, dedupe a CHANGELOG entry
Rebasing onto main pulled in #269/#286, which added
test_make_extractor_threads_the_real_source_into_calls_extraction to pin
mutmut mutant x__make_extractor__mutmut_10 (a source-argument mutation on
the extract_calls_generic call _make_extractor used to thread through).
This branch's own earlier commit deletes that exact call site and
extract_calls_generic itself (dead code, issue #249 boy-scout pass), so
the mutant the test existed to kill can no longer be generated from the
source at all -- git's line-based merge didn't conflict (the two edits
touch different lines), but the merged tree failed
test_ast_extractors_multilang.py with AttributeError on
ast_extractor_registry.extract_calls_generic. Root cause: a semantic
merge conflict between two independently-correct changes, invisible to
textual diffing. Fixed by deleting the retired test (its own in-file note
records why) rather than adapting it -- the 2-tuple _make_extractor now
composes exactly two calls, and a source-argument mutant on either is
already killed by the 7 per-language tests (test_java et al.), each of
which asserts on _imports/_defs output that is only correct when the REAL
source bytes reach imports_fn/defs_fn. Verified empirically, not just by
inspection: a scoped mutmut run (ast_extractor_registry.py +
ast_parser.py, full 6-file test selection) reports 50 "survived" mutants
recovered by mutation_recheck_survivors.py's full-selection reverification
(the same eager-import-dispatch-table false-positive pattern #269
diagnosed) -- 0 genuine survivors.

Boy-scout (seen while touching this tree): test_ast_parser.py had grown
to 405 lines under this branch's own Go/Swift/Rust wrapper-test additions,
over CLAUDE.md's 300-line/file cap -- a NEW violation this branch
introduces (main's version is 171 lines), not pre-existing debt. Split
(Extract Module, zero behavior change, same test names/assertions) into
test_ast_parser.py (223 lines), test_ast_parser_node_extraction.py (103
lines: the fake-Node _extract_module_doc/_node_text unit tests) and
test_ast_parser_extra_langs.py (122 lines: the Go/Swift/Rust
parse_file_ast integration tests) -- all three under cap.

Also seen: CHANGELOG.md's [Unreleased] section carried the #249
boy-scout entry TWICE (a stale pre-extract_calls_generic-deletion draft
alongside the final one), both stating a suite count already stale
relative to main's independent growth. Deduped to the single correct
entry; reworded its count claim to a relative delta (net +20 tests) per
the moment any other PR merges first -- assets/badge-tests.svg is the
one artifact that still states one).

Suite count re-measured post-fix: 6776 (pytest --collect-only -q on this
tree), matching the full run (6771 passed + 5 skipped). assets/badge-tests.svg
regenerated via scripts/generate_repo_badges.py --test-count 6776;
scripts/check_doc_claims.py --test-count 6776 and
scripts/generate_repo_badges.py --check --test-count 6776 both green.

Verification: tests_py/core/test_ast_parser.py +
test_ast_parser_node_extraction.py + test_ast_parser_extra_langs.py +
test_ast_extractors.py + test_ast_extractors_multilang.py +
test_ast_parser_language_contract.py (69 passed, unchanged assertions),
tests_py/core/ (3384 passed), full suite (6771 passed, 5 skipped, 116
subtests passed), ruff check + ruff format --check clean on all changed
files, pyright 0 errors/0 warnings on all changed files.
