+++ tests_py/core/test_ast_parser.py
@@ -3,6 +3,12 @@
 Tests adapt to the environment:
 - With tree-sitter: verifies full AST extraction (imports, classes, methods, etc.)
 - Without tree-sitter: verifies regex fallback produces valid FileAnalysis
+
+The fake-Node direct unit tests for `_extract_module_doc`/`_node_text`
+(`TestExtractModuleDoc`, `TestNodeText`, `_FakeNode`) live in the sibling
+`test_ast_parser_node_extraction.py` — split out to keep this file, which
+grew with the Go/Swift/Rust wrapper classes below, under the repo's
+300-line/file cap (CLAUDE.md, Code Style).
 """
 
 from __future__ import annotations
@@ -11,35 +17,9 @@ import sys
 
 import pytest
 
-from mcp_server.core.ast_parser import (
-    _extract_module_doc,
-    _node_text,
-    is_available,
-    parse_file_ast,
-)
+from mcp_server.core.ast_parser import is_available, parse_file_ast
 
 _HAS_TREE_SITTER = is_available()
-
-
-class _FakeNode:
-    """Minimal duck-typed stand-in for a tree_sitter.Node.
-
-    ast_parser's docstring/comment extraction only reads `.type`,
-    `.children`, `.start_byte`, `.end_byte` — a real grammar is not needed
-    to pin these functions' branches precisely.
-    """
-
-    def __init__(
-        self,
-        node_type: str,
-        children: list["_FakeNode"] | None = None,
-        start_byte: int = 0,
-        end_byte: int = 0,
-    ) -> None:
-        self.type = node_type
-        self.children = children or []
-        self.start_byte = start_byte
-        self.end_byte = end_byte
 
 
 class TestParseFilePython:
@@ -174,71 +154,6 @@ def test_is_available_false_when_tree_sitter_import_fails(
     assert is_available() is False
 
 
-class TestExtractModuleDoc:
-    """Direct unit tests for the docstring/comment-extraction branches —
-    fake nodes pin exact behavior without depending on a specific grammar's
-    tree shape (issue #249 mutation-testing pass)."""
-
-    def test_no_children_returns_empty(self) -> None:
-        root = _FakeNode("module", children=[])
-        assert _extract_module_doc(root, "python", b"") == ""
-
-    def test_python_bare_string_first_child(self) -> None:
-        source = b'"""hello"""'
-        string_node = _FakeNode("string", start_byte=0, end_byte=len(source))
-        root = _FakeNode("module", children=[string_node])
-        assert _extract_module_doc(root, "python", source) == "hello"
-
-    def test_python_expression_statement_wrapper(self) -> None:
-        source = b'"""Xylophone is great."""'
-        string_node = _FakeNode("string", start_byte=0, end_byte=len(source))
-        other = _FakeNode("other")
-        expr_stmt = _FakeNode("expression_statement", children=[string_node, other])
-        root = _FakeNode("module", children=[expr_stmt])
-        assert _extract_module_doc(root, "python", source) == "Xylophone is great."
-
-    def test_python_docstring_truncated_at_200(self) -> None:
-        body = "x" * 250
-        source = f'"""{body}"""'.encode()
-        string_node = _FakeNode("string", start_byte=0, end_byte=len(source))
-        root = _FakeNode("module", children=[string_node])
-        doc = _extract_module_doc(root, "python", source)
-        assert len(doc) == 200
-
-    def test_non_python_comment_first_child(self) -> None:
-        source = b"# Xray diagnostics."
-        comment = _FakeNode("comment", start_byte=0, end_byte=len(source))
-        root = _FakeNode("module", children=[comment])
-        assert _extract_module_doc(root, "javascript", source) == "Xray diagnostics."
-
-    def test_comment_truncated_at_200(self) -> None:
-        source = ("# " + "y" * 250).encode()
-        comment = _FakeNode("comment", start_byte=0, end_byte=len(source))
-        root = _FakeNode("module", children=[comment])
-        doc = _extract_module_doc(root, "javascript", source)
-        assert len(doc) == 200
-
-    def test_neither_string_nor_comment_returns_empty(self) -> None:
-        root = _FakeNode("module", children=[_FakeNode("class_declaration")])
-        assert _extract_module_doc(root, "go", b"") == ""
-
-
-class TestNodeText:
-    def test_extracts_slice_as_utf8(self) -> None:
-        source = b"hello world"
-        node = _FakeNode("identifier", start_byte=0, end_byte=5)
-        assert _node_text(node, source) == "hello"
-
-    def test_replaces_invalid_utf8_instead_of_raising(self) -> None:
-        """`errors="replace"` must survive a resolver swap, per issue #249:
-        an invalid error-handler name (a stub/version mismatch) raises
-        LookupError here instead of degrading gracefully."""
-        source = b"\xff\xfe"
-        node = _FakeNode("raw", start_byte=0, end_byte=len(source))
-        text = _node_text(node, source)
-        assert isinstance(text, str)
-
-
 class TestParseFileTypeScript:
     SAMPLE = b"""import { Request } from 'express';
 import jwt from 'jsonwebtoken';
@@ -302,104 +217,7 @@ class TestFallbackForUnsupported:
         assert r.definitions == []
 
 
-class TestParseFileGo:
-    """Exercises ast_parser._extract_go through the public parse_file_ast
-    entry point — TestGoExtractors (test_ast_extractors.py) calls
-    extract_go_definitions directly and never reaches this wrapper
-    (issue #249 mutation-testing pass)."""
-
-    SAMPLE = b"""package main
-
-import "fmt"
-
-type Server struct {
-    port int
-}
-
-func (s *Server) Start() error {
-    return nil
-}
-
-func NewServer(port int) *Server {
-    return &Server{port: port}
-}
-"""
-
-    def test_returns_file_analysis(self) -> None:
-        r = parse_file_ast("main.go", self.SAMPLE)
-        assert r.language == "go"
-
-    def test_imports(self) -> None:
-        r = parse_file_ast("main.go", self.SAMPLE)
-        if _HAS_TREE_SITTER:
-            assert "fmt" in [i.module for i in r.imports]
-
-    def test_definitions(self) -> None:
-        r = parse_file_ast("main.go", self.SAMPLE)
-        if _HAS_TREE_SITTER:
-            names = [d.name for d in r.definitions]
-            assert "Server.Start" in names
-            assert "NewServer" in names
-
-
-class TestParseFileSwift:
-    """Exercises ast_parser._extract_swift through parse_file_ast — no
-    prior test reached this wrapper at all (issue #249 mutation-testing
-    pass: 12/12 mutants reported "no tests" before this class)."""
-
-    SAMPLE = b"""import Foundation
-
-class AuthService {
-    func verify(token: String) -> Bool {
-        return true
-    }
-}
-"""
-
-    def test_returns_file_analysis(self) -> None:
-        r = parse_file_ast("Auth.swift", self.SAMPLE)
-        assert r.language == "swift"
-
-    def test_imports(self) -> None:
-        r = parse_file_ast("Auth.swift", self.SAMPLE)
-        if _HAS_TREE_SITTER:
-            assert "Foundation" in [i.module for i in r.imports]
-
-    def test_definitions(self) -> None:
-        r = parse_file_ast("Auth.swift", self.SAMPLE)
-        if _HAS_TREE_SITTER:
-            names = [d.name for d in r.definitions]
-            assert "AuthService" in names
-
-
-class TestParseFileRust:
-    """Exercises ast_parser._extract_rust through parse_file_ast — no
-    prior test reached this wrapper at all (issue #249 mutation-testing
-    pass: 12/12 mutants reported "no tests" before this class)."""
-
-    SAMPLE = b"""use std::fmt;
-
-struct Server {
-    port: u16,
-}
-
-fn new_server(port: u16) -> Server {
-    Server { port }
-}
-"""
-
-    def test_returns_file_analysis(self) -> None:
-        r = parse_file_ast("main.rs", self.SAMPLE)
-        assert r.language == "rust"
-
-    def test_imports(self) -> None:
-        r = parse_file_ast("main.rs", self.SAMPLE)
-        if _HAS_TREE_SITTER:
-            assert "std::fmt" in [i.module for i in r.imports]
-
-    def test_definitions(self) -> None:
-        r = parse_file_ast("main.rs", self.SAMPLE)
-        if _HAS_TREE_SITTER:
-            names = [d.name for d in r.definitions]
-            assert "Server" in names
-            assert "new_server" in names
+# TestParseFileGo/TestParseFileSwift/TestParseFileRust live in the sibling
+# test_ast_parser_extra_langs.py — split out to keep this file under the
+# repo's 300-line/file cap once those three classes were added (issue #249
+# mutation-testing pass).
